+++ x/dex/simulation/operations.go
@@ -13,6 +13,7 @@ import (
 	"github.com/cosmos/cosmos-sdk/x/simulation"
 
 	"github.com/NibiruChain/nibiru/x/common"
+	"github.com/NibiruChain/nibiru/x/common/denoms"
 	"github.com/NibiruChain/nibiru/x/dex/keeper"
 	"github.com/NibiruChain/nibiru/x/dex/types"
 )
@@ -364,9 +365,9 @@ func genPoolAssets(
 func fundAccountWithTokens(ctx sdk.Context, address sdk.AccAddress, bk types.BankKeeper) {
 	million := 1 * common.Precision
 	newTokens := sdk.NewCoins(
-		sdk.NewCoin(common.DenomNIBI, sdk.NewInt(int64(10*million))),
-		sdk.NewCoin(common.DenomUSDC, sdk.NewInt(int64(10*million))),
-		sdk.NewCoin(common.DenomNUSD, sdk.NewInt(int64(10*million))),
+		sdk.NewCoin(denoms.NIBI, sdk.NewInt(int64(10*million))),
+		sdk.NewCoin(denoms.USDC, sdk.NewInt(int64(10*million))),
+		sdk.NewCoin(denoms.NUSD, sdk.NewInt(int64(10*million))),
 	)
 
 	err := bk.MintCoins(ctx, types.ModuleName, newTokens)