+++ x/oracle/types/msgs_test.go
@@ -3,8 +3,7 @@ package types_test
 import (
 	"testing"
 
-	"github.com/NibiruChain/nibiru/x/common"
-
+	"github.com/NibiruChain/nibiru/x/common/denoms"
 	"github.com/NibiruChain/nibiru/x/oracle/types"
 
 	"github.com/stretchr/testify/require"
@@ -44,7 +43,7 @@ func TestMsgAggregateExchangeRatePrevote(t *testing.T) {
 		sdk.AccAddress([]byte("addr1_______________")),
 	}
 
-	exchangeRates := sdk.DecCoins{sdk.NewDecCoinFromDec(common.DenomUSDC, sdk.OneDec()), sdk.NewDecCoinFromDec(common.DenomNUSD, sdk.NewDecWithPrec(32121, 1))}
+	exchangeRates := sdk.DecCoins{sdk.NewDecCoinFromDec(denoms.USDC, sdk.OneDec()), sdk.NewDecCoinFromDec(denoms.NUSD, sdk.NewDecWithPrec(32121, 1))}
 	bz := types.GetAggregateVoteHash("1", exchangeRates.String(), sdk.ValAddress(addrs[0]))
 
 	tests := []struct {